need2fix/requirement.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.3
django-filter==1.1.0
git+https://github.com/scitu/django-tuauth.git
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
oauthlib==2.0.7
Pillow==5.0.0
pkg-resources==0.0.0
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
